streaming/requirements.txt
apache-beam[gcp]==2.50.0
google-cloud-aiplatform==1.35.0
